backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
pydantic[email]
passlib[bcrypt]
pytest
requests
pytest_cov==2.11.1
python-jose
python-multipart
